pyproject.toml
[tool.poetry]
name = "home_assignment"
version = "0.0.1"
description = "Home assignment"
authors = [

]
maintainers = [

]

readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
fastapi = "^0.78.0"
uvicorn = { version = "^0.17.6", extras = ["standard"] }
SQLAlchemy = {extras = ["mypy"], version = "^1.4.40"}
asyncpg = {version = "^0.26.0", extras = ["sa"]}
aiofiles = "^0.8.0"
httptools = "^0.4.0"
greenlet = "^1.1.2"
strawberry-graphql = {extras = ["fastapi"], version = "^0.125.0"}
httpx = "^0.23.0"
python-jose = "^3.3.0"
black = {extras = ["d"], version = "^22.8.0"}
yappi = "^1.3.6"
arrow = "^1.2.3"
passlib = "^1.7.4"
bcrypt = "^4.0.0"
SQLAlchemy-Utils = "^0.38.3"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
flake8 = "^4.0.1"
mypy = "^0.961"
isort = "^5.9.3"
yesqa = "^1.2.3"
wemake-python-styleguide = "^0.16.1"
black = {extras = ["d"], version = "^22.8.0"}
autoflake = "^1.4"
SQLAlchemy = {version = "^1.4", extras = ["mypy"]}
pytest-cov = "^3.0.0"
anyio = "^3.5.0"
pytest-env = "^0.6.2"
types-ujson = "^5.4.0"
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
